# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SIMS PPOB API",
    "version": "1.0.0",
    "description": "Dokumentasi API SIMS PPOB untuk Modul Membership"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/login": {
      "post": {
        "tags": ["1. Module Membership"],
        "summary": "Login untuk mendapatkan JWT Token",
        "description": "Login untuk mendapatkan token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "janedoe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "SecurePass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "Login Sukses"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 102
                    },
                    "message": {
                      "type": "string",
                      "example": "Parameter email tidak sesuai format"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 103
                    },
                    "message": {
                      "type": "string",
                      "example": "Username atau password salah"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
	"/api/transaction": {
    "post": {
      "tags": ["3. Module Transaction"],
      "summary": "Transaction Service",
      "description": "API ini digunakan untuk melakukan transaksi service PPOB. Menggunakan Bearer Token JWT.",
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "service_code": {
                  "type": "string",
                  "example": "PULSA",
                  "description": "Kode layanan yang dipilih untuk transaksi"
                }
              },
              "required": ["service_code"]
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Transaksi Berhasil",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 0
                  },
                  "message": {
                    "type": "string",
                    "example": "Transaksi berhasil"
                  },
                  "data": {
                    "type": "object",
                    "properties": {
                      "invoice_number": {
                        "type": "string",
                        "example": "INV17082023-001"
                      },
                      "service_code": {
                        "type": "string",
                        "example": "PLN_PRABAYAR"
                      },
                      "service_name": {
                        "type": "string",
                        "example": "PLN Prabayar"
                      },
                      "transaction_type": {
                        "type": "string",
                        "example": "PAYMENT"
                      },
                      "total_amount": {
                        "type": "integer",
                        "example": 10000
                      },
                      "created_on": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-17T10:10:10.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "400": {
          "description": "Bad Request (Service tidak ditemukan)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 102
                  },
                  "message": {
                    "type": "string",
                    "example": "Service atau Layanan tidak ditemukan"
                  },
                  "data": {
                    "type": "object",
                    "nullable": true,
                    "example": null
                  }
                }
              }
            }
          }
        },
        "401": {
          "description": "Unauthorized (Token tidak valid atau kadaluwarsa)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 108
                  },
                  "message": {
                    "type": "string",
                    "example": "Token tidak valid atau kadaluwarsa"
                  },
                  "data": {
                    "type": "object",
                    "nullable": true,
                    "example": null
                  }
                }
              }
            }
          }
        }
      }
    }
  	},
...
  }
}